move extension class to their own file
